chore: rename file-storage-emulator to bring in line with class name
